# openapi.yaml
openapi: 3.1.0
info:
  title: vitanota API
  version: 0.2.0
  description: |-
    教員向け BtoB SaaS「vitanota」の REST API 仕様。

    ## 認証
    全エンドポイントが Auth.js の database セッション戦略（SP-07）でログイン必須。
    Cookie `next-auth.session-token` を介してセッション検証される。

    ## パス設計
    - `/api/public/*` — CloudFront でキャッシュ可能（is_public=true のリソース）
    - `/api/private/*` — Cache-Control: private, no-store

    ## エラーコード
    - 400 VALIDATION_ERROR — Zod バリデーション失敗
    - 401 UNAUTHORIZED — セッション無効
    - 403 FORBIDDEN / TENANT_LOCKED — 権限不足・テナント停止
    - 404 JOURNAL_NOT_FOUND / TAG_NOT_FOUND — 所有者でない or 存在しない
    - 423 TENANT_LOCKED — テナント suspended
    - 500 INTERNAL_ERROR — 予期しないエラー
  contact:
    name: vitanota dev
servers:
  - url: https://dev.vitanota.example.com
    description: Development
  - url: https://vitanota.example.com
    description: Production
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: Auth.js セッション Cookie（database 戦略）
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: 入力が不正です
      required:
        - error
        - message
...
    BatonNote:
      type: object
      properties:
        id:
          type: string
        studentId:
          type: string
        authorUserId:
          type:
            - string
            - "null"
        noteDate:
          type: string
        sign:
          type:
            - string
            - "null"
          enum:
            - good
            - concern
            - null
        content:
          type:
            - string
            - "null"
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - studentId
        - authorUserId
        - noteDate
        - sign
        - content
        - createdAt
        - updatedAt
...
    UpdateBatonNoteInput:
      type: object
      properties:
        sign:
          type:
            - string
            - "null"
          enum:
            - good
            - concern
            - null
        content:
          type:
            - string
            - "null"
          maxLength: 500
...
    GradeMeetingTaskView:
      type: object
      properties:
        taskId:
          type: string
        title:
          type: string
        dueDate:
          type:
            - string
            - "null"
        status:
          type: string
          example: backlog
        categoryId:
          type: string
        assignees:
          type: array
          items:
            type: object
            properties:
              userId:
                type: string
              name:
                type:
                  - string
                  - "null"
            required:
              - userId
              - name
      required:
        - taskId
        - title
        - dueDate
        - status
        - categoryId
        - assignees
...
    GradeMeetingTaskResponse:
      type: object
      properties:
        task:
          $ref: "#/components/schemas/GradeMeetingTaskView"
      required:
        - task
    CreateGradeMeetingTaskInput:
      type: object
      properties:
        meetingId:
          type: string
        categoryId:
          type: string
        title:
          type: string
          minLength: 1
          maxLength: 15
        dueDate:
          type: string
          pattern: ^\d{4}-\d{2}-\d{2}$
      required:
        - meetingId
        - categoryId
        - title
    UnlinkGradeMeetingTaskInput:
      type: object
      properties:
        meetingId:
          type: string
        taskId:
          type: string
      required:
        - meetingId
        - taskId
...
paths:
...
  /api/baton-relay/notes/{id}:
    patch:
      summary: 自分の一言を編集
      tags:
        - Baton Relay
      security:
        - *a2
      parameters:
        - schema:
            type: string
          required: true
          name: id
          in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBatonNoteInput"
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  note:
                    $ref: "#/components/schemas/BatonNote"
                required:
                  - note
        "400":
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 未認証
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: 権限不足
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: リソースが見つからない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "423":
          description: テナント停止中
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: 自分の一言を削除
      tags:
        - Baton Relay
      security:
        - *a2
      parameters:
        - schema:
            type: string
          required: true
          name: id
          in: path
      responses:
        "204":
          description: 削除成功
        "400":
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 未認証
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: 権限不足
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: リソースが見つからない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "423":
          description: テナント停止中
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/grade-meeting/tasks:
    post:
      summary: 学年の「やること」を1つ起こす（実体は既存 tasks）
      description: クラスに紐づかない仕事（行事の準備・学年通信・保護者対応など）。TODO の仕組みを学年会の中に二重に作らず、既存 tasks
        に作って中間テーブルで会に紐付ける。担当・期限・完了はタスク側の仕組みをそのまま使い、タスクタブにも出る。
      tags:
        - GradeMeeting
      security:
        - *a2
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGradeMeetingTaskInput"
      responses:
        "200":
          description: 起こしたやること
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GradeMeetingTaskResponse"
        "400":
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 未認証
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: 権限不足
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: リソースが見つからない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "423":
          description: テナント停止中
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: 会から「やること」を外す（タスク本体は残す）
      description: 紐付けだけを外す。タスクタブで生きているものを学年会の画面から消させない。
      tags:
        - GradeMeeting
      security:
        - *a2
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnlinkGradeMeetingTaskInput"
      responses:
        "204":
          description: 外した
        "400":
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 未認証
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: 権限不足
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: リソースが見つからない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "423":
          description: テナント停止中
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...